requirements.txt
discord.py[voice]==2.7.1
PyYAML==6.0.3